+++ src/main/java/util/PickerStrategy.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 public interface PickerStrategy {
     int next(int upperBoundExclusive);
